pyproject.toml
# Copyright (c) Ansible Project
# GNU General Public License v3.0+ (see LICENSES/GPL-3.0-or-later.txt or https://www.gnu.org/licenses/gpl-3.0.txt)
# SPDX-License-Identifier: GPL-3.0-or-later

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "antsibull-docs-loader"
dynamic = [
    "version",
]
description = "Library for loading Ansible documentation"
license = "GPL-3.0-or-later AND BSD-2-Clause AND MIT AND PSF-2.0"
license-files = {globs=["LICENSES/*.txt"]}
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Ansible",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]
requires-python = ">=3.9"
dependencies = [
    "PyYAML",
]


[[project.authors]]
name = "Felix Fontein"
email = "felix@fontein.de"

[[project.maintainers]]
name = "Felix Fontein"
email = "felix@fontein.de"

[[project.maintainers]]
name = "Maxwell G"
email = "maxwell@gtmx.me"

[project.urls]
"Source code" = "https://github.com/ansible-community/antsibull-docs-loader"
"Code of Conduct" = "https://docs.ansible.com/projects/ansible/latest/community/code_of_conduct.html"
"Bug tracker" = "https://github.com/ansible-community/antsibull-docs-loader/issues"

[project.optional-dependencies]
codeqa = [
    "antsibull-changelog",
    "flake8 >= 6.0.0",
    "pylint >= 2.15.7",
    "reuse",
]
coverage = [
    "coverage[toml]",
]
formatters = [
    "black>=24",
    "isort",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-error-for-skips",
    "ansible-core @ https://github.com/ansible/ansible/archive/devel.tar.gz",
]
typing = [
    "mypy",
    "types-PyYAML",
    "typing-extensions",
]
dev = [
    # Used by nox sessions
    "antsibull-docs-loader[codeqa]",
    "antsibull-docs-loader[coverage]",
    "antsibull-docs-loader[formatters]",
    "antsibull-docs-loader[test]",
    "antsibull-docs-loader[typing]",
    # misc
    "nox",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "src/antsibull_docs_loader/__init__.py"

[tool.isort]
profile = "black"

[tool.coverage.report]
# https://coverage.readthedocs.io/en/latest/excluding.html#advanced-exclusion
# These should use single quotes in TOML, as they're regular expressions.
exclude_lines = [
    'pragma: no cover',
    'def __repr__',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
    'if TYPE_CHECKING:',
    'if t.TYPE_CHECKING:',
    'class .*\bProtocol\):',
    '@(abc\.)?abstractmethod',
]

[tool.coverage.paths]
source = [
    "src",
    "*/site-packages",
]

[tool.mypy]
mypy_path = "stubs/"

[[tool.mypy.overrides]]
module = [
    "ansible.constants",
    "pytest",
    "semantic_version",
]
ignore_missing_imports = true
